@@ -1,4 +1,4 @@
-class SetDefaultForAuthorAndText < ActiveRecord::Migration
+class SetDefaultForAuthorAndText < ActiveRecord::Migration[4.2]
   def up
     change_column_default(:comments, :author, "")
     change_column_default(:comments, :text, "")
