@@ -1,4 +1,4 @@
-class ChangeAuthorAndTextToNotNull < ActiveRecord::Migration
+class ChangeAuthorAndTextToNotNull < ActiveRecord::Migration[4.2]
   def change
     change_column_null(:comments, :author, false, "")
     change_column_null(:comments, :text, false, "")
